pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=2.31.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]
